workspaces/cursor-cli_claude-4.6-opus_20260321T030900Z/cmd/cli/examples/math-agent/requirements.txt
langchain
langchain-openai
langgraph
agentcube_sdk
pydantic
python-dotenv
